+++ src/components/FeedList.js
@@ -20,7 +20,7 @@ import MoreVertIcon from '@material-ui/icons/MoreVert';
 
 import { ChannelFixedID } from '../constants/index';
 import { setFeedReadStatus, openFeed, loadHistoryFeeds, markAllAsRead } from '../actions/index';
-import { withNamespaces } from 'react-i18next';
+import { withTranslation } from 'react-i18next';
 import GA from '../utils/GA';
 
 const mapStateToProps = state => {
@@ -134,6 +134,8 @@ class FeedList extends Component {
                     return r;
                 }, state.arrangedFeeds);
                 return new Map([...arrangedMap.entries()].sort((a, b) => a[0] - b[0]));
+            } else {
+                return new Map();
             }
         }
         if (props.currentChannelId !== state.lastChannelId || 
@@ -380,4 +382,4 @@ class FeedList extends Component {
     }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(withNamespaces()(FeedList)));
+export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(withTranslation()(FeedList)));