+++ src/components/Reader.js
@@ -23,7 +23,7 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
     return {
-        syncState: () => dispatch(syncState()),
+        syncState: isloadConfig => dispatch(syncState(isloadConfig)),
         setDefaultState: () => dispatch(setDefaultState()),
         setupBackgroundConnection: () => dispatch(setupBackgroundConnection()),
         setCurrentFeeds: () => dispatch(setCurrentFeeds()),
@@ -60,13 +60,14 @@ const useStyles = makeStyles(theme => ({
 const Reader = props => {
     const {setupBackgroundConnection, syncState, setDefaultState, setCurrentFeeds} = props;
     const [synced, setSyneced] = useState(false);
+    const [isDarkTheme, setIsDarkTheme] = useState(false);
 
     useEffect(() => {
         setupBackgroundConnection();
 
         let isSubscribed = true;
         async function syncDomainPerfix() {
-            await syncState();
+            await syncState(true);
             if (isSubscribed) {
                 setDefaultState();
                 setCurrentFeeds();
@@ -77,7 +78,20 @@ const Reader = props => {
         return () => isSubscribed = false;
     }, [syncState, setDefaultState, setCurrentFeeds, setupBackgroundConnection, setSyneced]);
 
-    const isDarkTheme = props.theme === "dark";
+    useEffect(() => {
+        const media = window.matchMedia('(prefers-color-scheme: dark)');
+        const systemThemeChange = e => {
+            setIsDarkTheme(e.matches);
+        }
+        if (props.theme === "system") {
+            media.addListener(systemThemeChange);
+            setIsDarkTheme(media.matches);
+        } else {
+            setIsDarkTheme(props.theme === "dark");
+        }
+        return () => media.removeListener(systemThemeChange);
+    }, [props.theme, setIsDarkTheme]);
+
     const classes = useStyles(props);
 
     return !synced ? null : (