+++ src/components/ReaderMessageBar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from "react-redux";
 import { withStyles } from '@material-ui/core/styles';
-import { withNamespaces } from 'react-i18next';
+import { withTranslation } from 'react-i18next';
 
 import { triggerAction, closeMessageBar } from '../actions/index';
 import Button from '@material-ui/core/Button';
@@ -93,4 +93,4 @@ readerMessageBar.propTypes = {
     config: PropTypes.object.isRequired,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(withNamespaces()(readerMessageBar)));
+export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(withTranslation()(readerMessageBar)));