+++ src/setupTests.js
@@ -16,9 +16,18 @@ global.chrome = chrome;
 jest.mock('Config');
 jest.mock('react-i18next', () => ({
     // this mock makes sure any components using the translate HoC receive the t function as a prop
-    withNamespaces: () => Component => {
+    withTranslation: () => Component => {
         Component.defaultProps = { ...Component.defaultProps, t: () => "" };
         return Component;
     },
 }));
 
+window.matchMedia = jest.fn().mockImplementation(query => {
+    return {
+        matches: false,
+        media: query,
+        onchange: null,
+        addListener: jest.fn(),
+        removeListener: jest.fn(),
+    };
+});